\name{footballpenaltiesBL-package}
\alias{footballpenaltiesBL-package}
\alias{footballpenaltiesBL}
\docType{package}
\title{
\packageTitle{footballpenaltiesBL}
}
\description{
\packageDescription{footballpenaltiesBL}
}
\author{
\packageAuthor{footballpenaltiesBL}

Maintainer: \packageMaintainer{footballpenaltiesBL}
}
\keyword{ package }
\examples{
data(penalties)
penaltyPrint("Illgner", exact = FALSE, pendat = penalties)
penaltyPlot("Illgner", exact = FALSE, pendat = penalties)
}
